Relax MSRV to 1.64 (#2)
* Relax MSRV to 1.64

The MSRV should probably track ndarray's which is currently 1.64. This
commit relaxes the msrv to do that.

* Raise MSRV from 1.64 to 1.65

While ndarray uses 1.64 for it's MSRV one of the dependencies in this
project is hashbrown has it's MSRV at 1.65 so we can't use 1.64. This
commit updates the MSRV to match the lowest that can be used 1.65
because of this.

* Fix clippy failure